+++ src/providers/completion.rs
@@ -52,7 +52,8 @@ pub async fn complete(
             .parent()
             .map(|p| p.to_string_lossy().into_owned());
         let home = std::env::var("HOME").ok();
-        let base_paths = completion_config.base_paths(&workspace_folders, &parent, &home);
+        let base_paths =
+            completion_config.base_paths(&workspace_folders, parent.as_ref(), home.as_ref());
 
         future::try_join_all(base_paths.iter().map(async |base_path| {
             generate_completions(