+++ src/providers/mod.rs
@@ -1,4 +1,5 @@
 mod completion;
+mod compute_tokens;
 mod definition;
 mod document_link;